Move test data definitions to separate swift file
