+++ generator/swift/Tests/MAVLinkTests/Testdata.swift
@@ -1,0 +1,33 @@
+//
+//  Testdata.swift
+//  MAVLink
+//
+//  Created by Max Odnovolyk on 10/31/16.
+//  Copyright © 2016 Build Apps. All rights reserved.
+//
+
+import XCTest
+@testable import MAVLink
+
+extension XCTestCase {
+    
+    /// Loads data from test tlog file
+    var testTlogData: Data {
+        let bundle = Bundle(for: MAVLinkTests.self)
+        let path = bundle.url(forResource: "flight", withExtension: "tlog")
+        return try! Data(contentsOf: path!)
+    }
+    
+    var testHeartbeatMessage: Heartbeat {
+        return Heartbeat(type: 6, autopilot: 8, baseMode: 0, customMode: 0, systemStatus: 0, mavlinkVersion: 3)
+    }
+    
+    var testHeartbeatData: Data {
+        return Data(bytes: [0xFE,0x09,0x00,0xFF,0x00,0x00,0x00,0x00,0x00,0x00,0x06,0x08,0x00,0x00,0x03,0xA1,0xDF])
+    }
+    
+    var testStatustextData: Data {
+        return Data(bytes: [0xFE,0x33,0x00,0x01,0x01,0xFD,0x01,0x41,0x50,0x4D,0x3A,0x43,0x6F,0x70,0x74,0x65,0x72,0x20,0x56,0x33,0x2E,0x34,0x2D,0x64,0x65,0x76,0x20,0x28,0x62,0x64,0x64,0x64,0x66,0x61,0x65,0x35,0x29,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x90,0x07])
+    }
+
+}
